Tweak styling of category pages to more closely match currently styling.
Lots of little things going on in this change!

I've re-written the Sass so that the colour-coding is automatically generated using a mixin,
similar to how it's done for the Shopping Guide sub-categories. This means that the colours are
stored in a single variable (in colors.scss) and can easily be adjusted, and we can also add simple
hover effects more easily (which I've done here).

I'm also using an extend placeholder for bordered elements (primarily images) to simplify the code,
and a placeholder for the icons as well. I've made some adjustments to the font sizing of headers
so they're more legible on mobiles, although this will invariably be tweaked further after testing.